+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Author in API: View all, create and delete
 - Cover in API: View all, create and delete
 - Notify in API: View all, create and delete
+- President in API: View all, create and delete
 
 ## [0.6.0] - 2020-12-07
 ### Added